Move genesis block and delegates test to application functional system tests
